# docs/api/openapi.yml
components:
  schemas:
...
    AuthtypesGettableObjects:
      properties:
        resource:
          $ref: '#/components/schemas/AuthtypesResource'
        selectors:
          items:
            type: string
          type: array
      required:
      - resource
      - selectors
      type: object
...
    AuthtypesPatchableObjects:
      properties:
        additions:
          items:
            $ref: '#/components/schemas/AuthtypesGettableObjects'
          nullable: true
          type: array
        deletions:
          items:
            $ref: '#/components/schemas/AuthtypesGettableObjects'
          nullable: true
          type: array
      required:
      - additions
      - deletions
      type: object
...
    AuthtypesResource:
      properties:
        name:
          type: string
        type:
          type: string
      required:
      - name
      - type
      type: object
...
    ErrorsJSON:
      properties:
        code:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorsResponseerroradditional'
          type: array
        message:
          type: string
        url:
          type: string
      required:
      - code
      - message
      type: object
    ErrorsResponseerroradditional:
      properties:
        message:
          type: string
      type: object
...
    RenderErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorsJSON'
        status:
          type: string
      required:
      - status
      - error
      type: object
...
  securitySchemes:
    api_key:
      description: API Keys
      in: header
      name: Hanzo O11y-Api-Key
      type: apiKey
    tokenizer:
      bearerFormat: Tokenizer
      description: Tokens generated by the tokenizer
      scheme: bearer
      type: http
info:
  description: OpenTelemetry-Native Logs, Metrics and Traces in a single pane
  title: Hanzo O11y
  version: ""
openapi: 3.0.3
paths:
...
  /api/v1/roles/{id}/relation/{relation}/objects:
    get:
      deprecated: false
      description: Gets all objects connected to the specified role via a given relation
        type
      operationId: GetObjects
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: relation
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: '#/components/schemas/AuthtypesGettableObjects'
                    type: array
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "451":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unavailable For Legal Reasons
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Implemented
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Get objects for a role by relation
      tags:
      - role
    patch:
      deprecated: false
      description: Patches the objects connected to the specified role via a given
        relation type
      operationId: PatchObjects
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: relation
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthtypesPatchableObjects'
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Found
        "451":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unavailable For Legal Reasons
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Not Implemented
      security:
      - api_key:
        - ADMIN
      - tokenizer:
        - ADMIN
      summary: Patch objects for a role by relation
      tags:
      - role
...